feat: implement loss functions
