+++ .gitignore
@@ -1,3 +1,5 @@
-/data/
 __pycache__/
 *.py[cod]
+
+/data/
+/wandb/
